# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sen Market API - Plateforme d'annonces pour le Sénégal
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  isAdmin: boolean;
  createdAt: string;
}

export interface RegisterInput {
  email: string;
  password: string;
  name: string;
  phone?: string;
  whatsapp?: string;
  city?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface UserUpdate {
  name?: string;
  phone?: string;
  whatsapp?: string;
  city?: string;
  avatarUrl?: string;
}

export interface Listing {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  price?: number | null;
  category: string;
  city: string;
  photos?: string[];
  status: string;
  userId: number;
  user?: User;
  createdAt: string;
  updatedAt: string;
}

export interface ListingInput {
  title: string;
  description: string;
  price?: number;
  category: string;
  city: string;
  photos?: string[];
}

export interface ListingUpdate {
  title?: string;
  description?: string;
  price?: number;
  category?: string;
  city?: string;
  photos?: string[];
  status?: string;
}

export interface ListingsPage {
  listings: Listing[];
  total: number;
  page: number;
  totalPages: number;
}

export interface CategoryStat {
  category: string;
  count: number;
}

export interface Message {
  id: number;
  content: string;
  senderId: number;
  receiverId: number;
  listingId: number;
  sender?: User;
  createdAt: string;
}

export interface MessageInput {
  content: string;
  receiverId: number;
  listingId: number;
}

export interface Conversation {
  listingId: number;
  listing?: Listing;
  otherUser: User;
  lastMessage?: Message;
  unreadCount: number;
}

export type AdminStatsListingsByCategoryItem = {
  category: string;
  count: number;
};

export interface AdminStats {
  totalUsers: number;
  totalListings: number;
  activeListings: number;
  pendingListings: number;
  rejectedListings: number;
  boostedListings: number;
  totalFavorites: number;
  totalReviews: number;
  listingsByCategory: AdminStatsListingsByCategoryItem[];
}

export interface Review {
  id: number;
  fromUserId: number;
  toUserId: number;
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
  fromUser?: User;
}

export interface ReviewInput {
  toUserId: number;
  rating: number;
  comment?: string;
}

export interface Notification {
  id: number;
  userId: number;
  type: string;
  message: string;
  isRead: boolean;
  /** @nullable */
  relatedId?: number | null;
  createdAt: string;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export type GetListingsParams = {
category?: string;
city?: string;
minPrice?: number;
maxPrice?: number;
search?: string;
page?: number;
limit?: number;
sortBy?: GetListingsSortBy;
sortOrder?: GetListingsSortOrder;
};

export type GetListingsSortBy = typeof GetListingsSortBy[keyof typeof GetListingsSortBy];


export const GetListingsSortBy = {
  date: 'date',
  price: 'price',
} as const;

export type GetListingsSortOrder = typeof GetListingsSortOrder[keyof typeof GetListingsSortOrder];


export const GetListingsSortOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;

export type AdminGetListingsParams = {
status?: string;
};