# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sen Market API - Plateforme d'annonces pour le Sénégal
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/listings:
    get:
      operationId: adminGetListings
      tags: [admin]
      summary: Get all listings for admin (all statuses)
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
        "403":
          description: Forbidden
  /admin/listings/{id}/approve:
    patch:
      operationId: adminApproveListing
      tags: [admin]
      summary: Approve a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "403":
          description: Forbidden
        "404":
          description: Not found
  /admin/listings/{id}/reject:
    patch:
      operationId: adminRejectListing
      tags: [admin]
      summary: Reject a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "403":
          description: Forbidden
        "404":
          description: Not found
  /admin/users:
    get:
      operationId: adminGetUsers
      tags: [admin]
      summary: Get all users for admin
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          description: Forbidden
  /admin/users/{id}/ban:
    patch:
      operationId: adminBanUser
      tags: [admin]
      summary: Ban a user (delete all their listings)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "403":
          description: Forbidden
        "404":
          description: Not found
...
  /favorites:
    get:
      operationId: getFavorites
      tags: [favorites]
      summary: Get current user's favorite listings
      security:
        - cookieAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
        "401":
          description: Unauthorized
  /favorites/{listingId}:
    post:
      operationId: addFavorite
      tags: [favorites]
      summary: Add a listing to favorites
      security:
        - cookieAuth: []
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Unauthorized
    delete:
      operationId: removeFavorite
      tags: [favorites]
      summary: Remove a listing from favorites
      security:
        - cookieAuth: []
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Unauthorized
...
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get global platform statistics
      security:
        - cookieAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "403":
          description: Forbidden
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        phone:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isAdmin:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - isAdmin
        - createdAt
...
    Listing:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        price:
          type: ["number", "null"]
        category:
          type: string
        city:
          type: string
        photos:
          type: array
          items:
            type: string
        status:
          type: string
        userId:
          type: integer
        user:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - description
        - category
        - city
        - status
        - userId
        - createdAt
        - updatedAt
...
    AdminStats:
      type: object
      properties:
        totalUsers:
          type: integer
        totalListings:
          type: integer
        activeListings:
          type: integer
        pendingListings:
          type: integer
        rejectedListings:
          type: integer
        boostedListings:
          type: integer
        totalFavorites:
          type: integer
        totalReviews:
          type: integer
        listingsByCategory:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              count:
                type: integer
            required: [category, count]
      required:
        - totalUsers
        - totalListings
        - activeListings
        - pendingListings
        - rejectedListings
        - boostedListings
        - totalFavorites
        - totalReviews
        - listingsByCategory
...