# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sen Market API - Plateforme d'annonces pour le Sénégal
 * OpenAPI spec version: 0.1.0
 */
import type { AdminStatsListingsByCategoryItem } from './adminStatsListingsByCategoryItem';

export interface AdminStats {
  totalUsers: number;
  totalListings: number;
  activeListings: number;
  pendingListings: number;
  rejectedListings: number;
  boostedListings: number;
  totalFavorites: number;
  totalReviews: number;
  listingsByCategory: AdminStatsListingsByCategoryItem[];
}